Better PasswordConfirm Form - super user deleting member account still impossible
The password check seems to be working so far, I chosed not to use the `@confirmpassword` because I wanted a special page with a red button delete.
A user can  delete it's own account but a super user going can't delete a member account, it's his own account that is proposed to be deleted, that's beacuse I didn't manage to understand how to get the delete_user (which is meant to be the previous edit_user), I can only manage to get the user which is logged in by `self.request.user`, I tried to add a `get_context_data` as done in `ProfileView`, it was just ignored...
Also I didn't put the Danger zone in css or modified the danger sign size yet.