Remove Marauder
It has not been used in the past few years, and the GPS was kind of
imprecise so it did not work very well. The lack of multi-device
compatibility was also a problem.

Right now it's a maintainance burden, so it's better to remove it
altogether. If someone wants to revive it, there will be a lot of work
required anyway, so it shouldn't be too hard to restore it from the git
history by reverting this commit.